fixed duplicate testing folder
